Introduce IRequestHandler.hpp
